Add Unauthorized check

+++ WishList/Controllers/ItemController.cs
@@ -35,7 +35,8 @@ namespace WishList.Controllers
         [HttpPost]
         public IActionResult Create(Models.Item item)
         {
-            item.User = _userManager.GetUserAsync(HttpContext.User).Result;
+            var user = _userManager.GetUserAsync(HttpContext.User).Result;
+            item.User = user;
             _context.Items.Add(item);
             _context.SaveChanges();
             return RedirectToAction("Index");
@@ -43,8 +44,11 @@ namespace WishList.Controllers
 
         public IActionResult Delete(int id)
         {
+            var user = _userManager.GetUserAsync(HttpContext.User).Result;
             var item = _context.Items.FirstOrDefault(e => e.Id == id);
-            item.User = _userManager.GetUserAsync(HttpContext.User).Result;
+            if (user != null)
+                return Unauthorized();
+            item.User = user;
             _context.Items.Remove(item);
             _context.SaveChanges();
             return RedirectToAction("Index");
